Velocity advection and pressure gradient subtraction
Using trirop's non-iterative Jacobi pressure solver as a ShaderToy hack since ShaderToy cannot iterate a single shader pass. It is based on running a Jacobi pressure solver for 20 passes and manually precalculating a 20x7 convolution kernel to apply to the image in one pass.

For the non-ShaderToy version we will use Gauss-Seidel method iterated normally.